@@ -4,15 +4,16 @@
   "private": true,
   "dependencies": {
     "history": "^4.7.2",
-    "react": "^15.0.0",
-    "react-dom": "^15.0.0",
-    "react-router": "^3.2.1",
+    "react": "^15.6.0",
+    "react-dom": "^15.6.0",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4",
     "redux-router": "^1.0.0-beta3",
     "router": "^1.3.3"
   },
   "scripts": {
-     "server": "cd server && json-server db.json -w -p 3001",  
+    "server": "cd server && json-server db.json -w -p 3001",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
