@@ -1,19 +1,26 @@
 import React, { Component } from 'react';
-import logo from './logo.svg';
 import './App.css';
+import { Route }  from 'react-router-dom';
+import HomePage from './pages/Home';
+import UserListPage from './pages/UserList';
+import UserEditPage from './pages/UserEdit';
+import { BrowserRouter } from 'react-router-dom';
+import UserAddPage from './pages/UserAdd';
+
+
 
 class App extends Component {
   render() {
     return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1 className="App-title">Welcome to React</h1>
-        </header>
-        <p className="App-intro">
-          To get started, edit <code>src/App.js</code> and save to reload.
-        </p>
-      </div>
+      <BrowserRouter>
+        <div>
+            <Route path="/" exact component = {HomePage} />   
+            <Route path="/user/add" exact component={UserAddPage}/>
+            <Route  path="/user/list" exact component = {UserListPage}/>
+            <Route  path="/user/edit/:id" exact component = {UserEditPage}/>
+          </div>
+  </BrowserRouter>
+
     );
   }
 }
