@@ -27,6 +27,24 @@ function formProvider (fields) {
                 //     this.handleValueChange = this.handleValueChange.bind(this);
                 // }
 
+                setFormValues = (values) => {
+                    if(!values) {
+                        return;
+                    }
+
+                    const {form} = this.state;
+                    let newForm = {...form};
+                    for(const field in form) {
+                        if(form.hasOwnProperty(field)) {
+                            if(typeof values[field] !== 'undefined') {
+                                newForm[field] = {...newForm[field], value: values[field]};
+                            }
+                            newForm[field].valid = true;
+                        }
+                    }
+                    this.setState({form:newForm});
+                }
+
                 handleValueChange  = (fieldName, value) => {
                     const { form } = this.state;
                     const newFieldState = {value, valid:true, error: ''};
@@ -60,8 +78,10 @@ function formProvider (fields) {
 
               render () {
                   const {form, formValid} = this.state;
-                  return <Comp {...this.props} form={form} formValid = {formValid} 
-                                onFormChange = {this.handleValueChange} />    //comp here as input, have additional property, form, formvalid, onformchange
+                  return (<Comp 
+                  {...this.props} form={form} formValid = {formValid} 
+                    onFormChange = {this.handleValueChange} 
+                  setFormValues = {this.setFormValues} />);   //comp here as input, have additional property, form, formvalid, onformchange
               }  
         }
     return FormComponent;  
